# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "USER - ENTITY API",
    "description": "USER AND ENTITY CRUD",
    "contact": {
      "email": "jtolojanahary5@gmail.com"
    },
    "license": {
      "name": "Portfolio",
      "url": "https://j-lucien.github.io/portfolio-angular/"
    },
    "version": "1.0.11"
  },
...
  "paths": {
    "/entities": {
      "post": {
        "tags": [
          "entities"
        ],
        "summary": "Create an entity",
        "description": "Create a new entity.",
        "operationId": "entityId",
        "requestBody": {
          "description": "Created entity object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/EntityRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/EntityRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/EntityResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "Get all entity",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/EntityResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/user-entities/{id}": {
      "put": {
        "tags": [
          "user-entities"
        ],
        "summary": "Update an existing user-entities",
        "description": "Update an existing user-entities by Id",
        "operationId": "updateUserEntity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Update an existent user-entities",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserEntityRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/UserEntityRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/UserEntityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEntity"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/UserEntity"
                }
              }
            }
          },
          "404": {
            "description": "Id not found"
          }
        }
      },
      "get": {
        "tags": [
          "user-entities"
        ],
        "summary": "Get an user-entities",
        "description": "Get an user-entities by Id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user-entities to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEntityResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/UserEntityResponse"
                }
              }
            }
          },
          "404": {
            "description": "Id not found"
          }
        }
      },
      "delete": {
        "tags": [
          "user-entities"
        ],
        "summary": "Delete user-entities",
        "operationId": "deleteUserEntity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id that needs to be deleted",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "user-entities supprimé avec succès !"
          },
          "404": {
            "description": "id not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Entity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Nom de l'entité",
            "maxLength": 50,
            "minLength": 1,
            "example": "My Entity"
          },
          "description": {
            "type": "string",
            "description": "Description détaillée de l'entité",
            "example": "This is a description of the entity."
          },
          "siret": {
            "type": "string",
            "description": "Numéro de SIRET de l'entité",
            "maxLength": 20,
            "example": "12345678901234"
          },
          "keyLicence": {
            "type": "string",
            "description": "Clé de licence associée à l'entité",
            "maxLength": 250,
            "example": "ABCD-1234-XYZ"
          },
          "website": {
            "type": "string",
            "description": "Site web de l'entité",
            "maxLength": 100,
            "example": "https://www.entitywebsite.com"
          }
        }
      },
      "EntityRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nom de l'entité",
            "maxLength": 50,
            "minLength": 1,
            "example": "My Entity"
          },
          "description": {
            "type": "string",
            "description": "Description détaillée de l'entité",
            "example": "This is a description of the entity."
          },
          "siret": {
            "type": "string",
            "description": "Numéro de SIRET de l'entité",
            "maxLength": 20,
            "example": "12345678901234"
          },
          "keyLicence": {
            "type": "string",
            "description": "Clé de licence associée à l'entité",
            "maxLength": 250,
            "example": "ABCD-1234-XYZ"
          },
          "website": {
            "type": "string",
            "description": "Site web de l'entité",
            "maxLength": 100,
            "example": "https://www.entitywebsite.com"
          }
        }
      },
...
      "EntityResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Nom de l'entité",
            "maxLength": 50,
            "minLength": 1,
            "example": "My Entity"
          },
          "description": {
            "type": "string",
            "description": "Description détaillée de l'entité",
            "example": "This is a description of the entity."
          },
          "siret": {
            "type": "string",
            "description": "Numéro de SIRET de l'entité",
            "maxLength": 20,
            "example": "12345678901234"
          },
          "keyLicence": {
            "type": "string",
            "description": "Clé de licence associée à l'entité",
            "maxLength": 250,
            "example": "ABCD-1234-XYZ"
          },
          "website": {
            "type": "string",
            "description": "Site web de l'entité",
            "maxLength": 100,
            "example": "https://www.entitywebsite.com"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date de création de l'entité",
            "example": "2024-11-10T12:34:56Z"
          },
          "Users": {
            "type": "array",
            "description": "Liste des utilisateurs associés à l'entité",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "example": 3
                },
                "name": {
                  "type": "string",
                  "example": "Jean Lucien"
                },
                "firstName": {
                  "type": "string",
                  "example": "Lucien"
                },
                "language": {
                  "type": "string",
                  "example": "fr"
                },
                "email": {
                  "type": "string",
                  "format": "email",
                  "example": "lucien@gmail.com"
                },
                "lastLogin": {
                  "type": "string",
                  "format": "date-time",
                  "nullable": true,
                  "example": null
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time",
                  "example": "2024-11-10T10:51:41.000Z"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time",
                  "example": "2024-11-10T10:51:41.000Z"
                }
              }
            }
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 3
          },
          "name": {
            "type": "string",
            "example": "Jean Lucien"
          },
          "firstName": {
            "type": "string",
            "example": "Lucien"
          },
          "language": {
            "type": "string",
            "example": "fr"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "lucien@gmail.com"
          }
        }
      },
...
      "UserEntityRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "example": 3
          },
          "entityId": {
            "type": "integer",
            "example": 1
          }
        }
      },
      "UserEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "userId": {
            "type": "integer",
            "example": 3
          },
          "entityId": {
            "type": "integer",
            "example": 1
          }
        }
      },
      "UserEntityResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "userId": {
            "type": "integer",
            "example": 3
          },
          "entityId": {
            "type": "integer",
            "example": 1
          },
          "User": {
            "$ref": "#/components/schemas/User"
          },
          "Entity": {
            "$ref": "#/components/schemas/Entity"
          }
        }
      },
...
    }
  }
}